feat: move protocol helper logic in procotol.rs
